requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1

# LLM & RAG
groq==0.9.0
langchain==0.2.1
langchain-community==0.2.1
langchain-groq==0.1.6
langchain-huggingface==0.0.3
langgraph==0.1.1
# Embeddings & Vector Store
faiss-cpu==1.13.2
sentence-transformers==3.0.0

# Document Processing
pypdf==4.2.0
tiktoken==0.7.0

# Utilities
aiofiles==23.2.1
httpx==0.27.0
numpy==1.26.4
